Move from resources to provider/cloud/kubevirt
