+++ pkg/provider/cloud/kubevirt/manifests/instancetype/kubermatic-standard.yaml
