+++ pkg/provider/cloud/kubevirt/manifests/manifests_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package kubevirtmanifests
+package manifests
 
 import (
 	"embed"
