+++ pkg/provider/cloud/kubevirt/manifests/preset/kubermatic-standard.yaml
