+++ pkg/provider/cloud/kubevirt/manifests/test/instancetype/kubermatic-standard.yaml
