+++ pkg/provider/cloud/kubevirt/manifests/test/preference/kubermatic-standard.yaml
