+++ pkg/provider/cloud/kubevirt/manifests/test/presetOk/kubermatic-standard-1.yaml
