+++ pkg/provider/cloud/kubevirt/manifests/test/presetOk/kubermatic-standard-2.yaml
