+++ pkg/provider/cloud/kubevirt/manifests/test/presetOkKo/kubermatic-standard-1.yaml
