+++ pkg/provider/cloud/kubevirt/manifests/test/presetOkKo/kubermatic-standard-2.yaml
